move bundle folder
